Refactor time functions
