+++ lib/modules/betteReddit.js
@@ -20,7 +20,7 @@ import {
 	watchForElements,
 } from '../utils';
 import { _addHeaderId } from '../utils/dom';
-import { fromYoutubeTimecodeToSeconds, fromSecondsToTime } from '../utils/time';
+import { fromSecondsToTime } from '../utils/time';
 import { ajax, i18n } from '../environment';
 import { Expando } from './showImages/expando';
 import { isSettingsUrl } from './settingsNavigation';
@@ -404,6 +404,20 @@ const getVideoInfo = batch(async videoIds => {
 
 	return videoIds.map(idFromBatch => results.find(({ id }) => id === idFromBatch));
 }, { size: 50 });
+
+function fromYoutubeTimecodeToSeconds(tc: string) {
+	let timeSeconds = Number(tc);
+	if (Number.isNaN(timeSeconds)) {
+		const tcobj = (tc.split(/(\d+[hms])/)
+			.filter(Boolean)
+			.reduce((acc, match) => {
+				acc[match.slice(-1)] = Number(match.slice(0, -1));
+				return acc;
+			}, {}));
+		timeSeconds = (tcobj.h || 0) * 3600 + (tcobj.m || 0) * 60 + (tcobj.s || 0);
+	}
+	return timeSeconds;
+}
 
 function pinSubredditBar() {
 	// Make the subreddit bar at the top of the page a fixed element
