+++ lib/utils/time.js
@@ -6,33 +6,14 @@ export const DAY = 24 * HOUR;
 export const WEEK = 7 * DAY;
 
 export function fromSecondsToTime(timeInSeconds: number) {
+	const seconds = timeInSeconds % 60;
+	const minutes = Math.floor(timeInSeconds / 60) % 60;
 	const hours = Math.floor(timeInSeconds / 3600);
-	const minutes = Math.floor((timeInSeconds - (hours * 3600)) / 60);
-	const seconds = timeInSeconds - (hours * 3600) - (minutes * 60);
-	let time = '';
 
-	if (hours !== 0) {
-		time += String(hours).padStart(2, '0');
-		time += ':';
+	const time = [minutes, seconds];
+	if (hours > 0) {
+		time.unshift(hours);
 	}
 
-	time += String(minutes).padStart(2, '0');
-	time += ':';
-	time += String(seconds).padStart(2, '0');
-
-	return time;
-}
-
-export function fromYoutubeTimecodeToSeconds(tc: string) {
-	let timeSeconds = Number(tc);
-	if (Number.isNaN(timeSeconds)) {
-		const tcobj = (tc.split(/(\d+[hms])/)
-			.filter(Boolean)
-			.reduce((acc, match) => {
-				acc[match.slice(-1)] = Number(match.slice(0, -1));
-				return acc;
-			}, {}));
-		timeSeconds = (tcobj.h || 0) * 3600 + (tcobj.m || 0) * 60 + (tcobj.s || 0);
-	}
-	return timeSeconds;
+	return time.map(part => String(part).padStart(2, '0')).join(':');
 }
